+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `context` parameter to `DataCube.aggregate_spatial()`, `DataCube.apply_dimension()`, 
   `DataCube.apply_neighborhood()`, `DataCube.apply()`, `DataCube.merge_cubes()`.
   ([#291](https://github.com/Open-EO/openeo-python-client/issues/291))
+- Add `DataCube.fit_regr_random_forest()` ([#293](https://github.com/Open-EO/openeo-python-client/issues/293))
 
 ### Changed